# gen/openapi/hub.openapi.yaml
openapi: 3.1.0
info:
  title: gravel hub
  description: The HTTP-JSON side of the gravel hub's Connect API. The same procedures serve gRPC and gRPC-Web.
  version: v1
  license:
    name: Apache-2.0
    identifier: Apache-2.0
paths:
...
  /gravel.hub.v1.OrganizationService/GetOrganizationSettings:
    post:
      tags:
        - gravel.hub.v1.OrganizationService
      summary: GetOrganizationSettings
      description: |-
        GetOrganizationSettings returns the settings; public, because the pages render the theme
         before anyone logs in, and the document holds no credential.
      operationId: gravel.hub.v1.OrganizationService.GetOrganizationSettings
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.GetOrganizationSettingsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.GetOrganizationSettingsResponse'
...
  /gravel.hub.v1.ServerService/GetServerStatus:
    post:
      tags:
        - gravel.hub.v1.ServerService
      summary: GetServerStatus
      description: GetServerStatus is one server; public. not_found for an unknown id.
      operationId: gravel.hub.v1.ServerService.GetServerStatus
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.GetServerStatusRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.GetServerStatusResponse'
  /gravel.hub.v1.ServerService/ListGames:
    post:
      tags:
        - gravel.hub.v1.ServerService
      summary: ListGames
      description: ListGames is the enabled games; public.
      operationId: gravel.hub.v1.ServerService.ListGames
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.ListGamesRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.ListGamesResponse'
  /gravel.hub.v1.ServerService/ListServerPlayers:
    post:
      tags:
        - gravel.hub.v1.ServerService
      summary: ListServerPlayers
      description: |-
        ListServerPlayers is who was on at the last poll: names and provider identities, so a
         logged-in member, the owner, or an app with the servers:read scope.
      operationId: gravel.hub.v1.ServerService.ListServerPlayers
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.ListServerPlayersRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.ListServerPlayersResponse'
  /gravel.hub.v1.ServerService/ListServers:
    post:
      tags:
        - gravel.hub.v1.ServerService
      summary: ListServers
      description: ListServers is every server with its last observation; public, no personal data.
      operationId: gravel.hub.v1.ServerService.ListServers
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.ListServersRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.ListServersResponse'
  /gravel.hub.v1.StatsService/GetBoard:
    post:
      tags:
        - gravel.hub.v1.StatsService
      summary: GetBoard
      description: |-
        GetBoard ranks the players of a server, a game or the organization over a window. Public when
         the Organization settings make stats public; otherwise the owner or an app with stats:read.
      operationId: gravel.hub.v1.StatsService.GetBoard
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.GetBoardRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.GetBoardResponse'
...
  /gravel.hub.v1.StatsService/ListMatches:
    post:
      tags:
        - gravel.hub.v1.StatsService
      summary: ListMatches
      description: ListMatches is a server's matches, newest first; the same access as GetBoard.
      operationId: gravel.hub.v1.StatsService.ListMatches
      parameters:
        - name: Connect-Protocol-Version
          in: header
          description: Define the version of the Connect protocol
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
          example: 1
        - name: Connect-Timeout-Ms
          in: header
          description: Define the timeout, in ms
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
          example: 1000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/gravel.hub.v1.ListMatchesRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/gravel.hub.v1.ListMatchesResponse'
...
components:
  schemas:
    connect-protocol-version:
      type: number
      examples:
        - 1
      title: Connect-Protocol-Version
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      examples:
        - 1000
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
            - not_found
          enum:
            - canceled
            - unknown
            - invalid_argument
            - deadline_exceeded
            - not_found
            - already_exists
            - permission_denied
            - resource_exhausted
            - failed_precondition
            - aborted
            - out_of_range
            - unimplemented
            - internal
            - unavailable
            - data_loss
            - unauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/connect.error_details.Any'
          description: A list of messages that carry the error details. There is no limit on the number of messages.
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
    connect.error_details.Any:
      type: object
      properties:
        type:
          type: string
          description: 'A URL that acts as a globally unique identifier for the type of the serialized message. For example: `type.googleapis.com/google.rpc.ErrorInfo`.'
        value:
          type: string
          format: byte
          description: The Protobuf message, serialized as bytes and base64-encoded. The specific message type is identified by the `type` field.
        debug:
          type: object
          title: Debug
          additionalProperties: true
          description: Deserialized error detail payload. The 'type' field indicates the schema. This field is for easier debugging and should not be relied upon for application logic.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message, with an additional debug field for ConnectRPC error details.
...
    google.protobuf.Timestamp:
      type: string
      examples:
        - "2023-01-15T01:30:15.01Z"
        - "2024-12-25T12:00:00Z"
      format: date-time
      description: A point in time in RFC 3339 format, with up to nanosecond precision. Output uses UTC (`Z`); input may use an offset from UTC.
...
    gravel.hub.v1.BoardEntry:
      type: object
      properties:
        rank:
          type: integer
          title: rank
          format: int32
        name:
          type: string
          title: name
        pseudonymous:
          type: boolean
          title: pseudonymous
        userId:
          type: string
          title: user_id
          description: The member, when shown by name.
        kills:
          type: integer
          title: kills
          format: int32
        deaths:
          type: integer
          title: deaths
          format: int32
        kd:
          type: number
          title: kd
          format: double
          description: kills / max(deaths, 1).
        secondsOn:
          type: integer
          title: seconds_on
          format: int32
        matches:
          type: integer
          title: matches
          format: int32
      title: BoardEntry
      additionalProperties: false
      description: |-
        BoardEntry is one player on a board. A player who has not linked the identity, or has not
         chosen to be shown by name, appears under their pseudonym (pseudonymous true, no user_id).
...
    gravel.hub.v1.Game:
      type: object
      properties:
        id:
          type: string
          title: id
          description: '"wardogs".'
        name:
          type: string
          title: name
        identityProvider:
          type: string
          title: identity_provider
          description: 'The provider its players are keyed by: "steam".'
        bands:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.Band'
          title: bands
        noPaidPerks:
          type: boolean
          title: no_paid_perks
          description: |-
            The publisher forbids any paid benefit (a supporter perk, a crowdfunding reward) that depends
             on the game's data or features.
        layout:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.GameLayout'
          title: layout
          description: How the pages show the game (the Game spec's layout).
      title: Game
      additionalProperties: false
      description: |-
        Game is a game this hub's servers run: a Game spec gravel ships, with the deployment's
         tightenings of its bands applied.
...
    gravel.hub.v1.GetBoardRequest:
      type: object
      properties:
        serverId:
          type: string
          title: server_id
          description: |-
            At most one of server_id and game_id; neither is the organization's board. A server's board
             counts its matches whatever its trust; a game's or the organization's counts Official
             servers' only (ADR-0011's trust mapping).
        gameId:
          type: string
          title: game_id
        window:
          type: string
          title: window
          description: '"week" (this week), "month" (this month), "season" (the named one), "all" (the default).'
        season:
          type: string
          title: season
        metric:
          type: string
          title: metric
          description: '"kills" (the default), "deaths", "kd", "time", "matches".'
        pageSize:
          type: integer
          title: page_size
          format: int32
        pageToken:
          type: string
          title: page_token
          description: The next_page_token of the previous page; empty for the first.
      title: GetBoardRequest
      additionalProperties: false
    gravel.hub.v1.GetBoardResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.BoardEntry'
          title: entries
        nextPageToken:
          type: string
          title: next_page_token
        from:
          allOf:
            - $ref: '#/components/schemas/google.protobuf.Timestamp'
          title: from
          description: The window's bounds; unset for "all".
        to:
          allOf:
            - $ref: '#/components/schemas/google.protobuf.Timestamp'
          title: to
      title: GetBoardResponse
      additionalProperties: false
...
    gravel.hub.v1.GetOrganizationSettingsRequest:
      type: object
      title: GetOrganizationSettingsRequest
      additionalProperties: false
    gravel.hub.v1.GetOrganizationSettingsResponse:
      type: object
      properties:
        settings:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.OrganizationSettings'
          title: settings
      title: GetOrganizationSettingsResponse
      additionalProperties: false
...
    gravel.hub.v1.GetServerStatusRequest:
      type: object
      properties:
        serverId:
          type: string
          title: server_id
      title: GetServerStatusRequest
      additionalProperties: false
    gravel.hub.v1.GetServerStatusResponse:
      type: object
      properties:
        server:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.Server'
          title: server
      title: GetServerStatusResponse
      additionalProperties: false
...
    gravel.hub.v1.ListGamesRequest:
      type: object
      title: ListGamesRequest
      additionalProperties: false
    gravel.hub.v1.ListGamesResponse:
      type: object
      properties:
        games:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.Game'
          title: games
      title: ListGamesResponse
      additionalProperties: false
...
    gravel.hub.v1.ListMatchesRequest:
      type: object
      properties:
        serverId:
          type: string
          title: server_id
        pageSize:
          type: integer
          title: page_size
          format: int32
        pageToken:
          type: string
          title: page_token
          description: The next_page_token of the previous page; empty for the first.
      title: ListMatchesRequest
      additionalProperties: false
    gravel.hub.v1.ListMatchesResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.MatchSummary'
          title: matches
        nextPageToken:
          type: string
          title: next_page_token
      title: ListMatchesResponse
      additionalProperties: false
...
    gravel.hub.v1.ListServerPlayersRequest:
      type: object
      properties:
        serverId:
          type: string
          title: server_id
      title: ListServerPlayersRequest
      additionalProperties: false
    gravel.hub.v1.ListServerPlayersResponse:
      type: object
      properties:
        players:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.ServerPlayer'
          title: players
        observedAt:
          allOf:
            - $ref: '#/components/schemas/google.protobuf.Timestamp'
          title: observed_at
          description: The poll the list is from; unset when there was none.
      title: ListServerPlayersResponse
      additionalProperties: false
    gravel.hub.v1.ListServersRequest:
      type: object
      title: ListServersRequest
      additionalProperties: false
    gravel.hub.v1.ListServersResponse:
      type: object
      properties:
        servers:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.Server'
          title: servers
      title: ListServersResponse
      additionalProperties: false
...
    gravel.hub.v1.OrganizationSettings:
      type: object
      properties:
        version:
          type: integer
          title: version
          format: int32
        theme:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.Theme'
          title: theme
        nav:
          type: array
          items:
            $ref: '#/components/schemas/gravel.hub.v1.NavLink'
          title: nav
        updatedAt:
          allOf:
            - $ref: '#/components/schemas/google.protobuf.Timestamp'
          title: updated_at
          description: When the settings were last written; unset while the defaults apply.
        discord:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.DiscordSettings'
          title: discord
        stats:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.StatsSettings'
          title: stats
      title: OrganizationSettings
      additionalProperties: false
      description: |-
        OrganizationSettings is the Organization settings resource: every host-configurable knob in
         one document (theme and navigation, the Discord role mapping; token lifetimes and layout
         overrides join it with their features). Version is the document format.
...
    gravel.hub.v1.Server:
      type: object
      properties:
        id:
          type: string
          title: id
          description: '"htg-wardogs-1".'
        name:
          type: string
          title: name
        gameId:
          type: string
          title: game_id
        location:
          type: string
          title: location
          description: 'Where it runs: "qonzer-slc".'
        trust:
          type: string
          title: trust
          description: '"official" or "community".'
        seeding:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.Seeding'
          title: seeding
          description: Unset when the server does not seed.
        status:
          allOf:
            - $ref: '#/components/schemas/gravel.hub.v1.ServerStatus'
          title: status
      title: Server
      additionalProperties: false
      description: |-
        Server is a server a driver controls (ManagedServer), as the public sees it: never its control
         address or where its credential is.
    gravel.hub.v1.ServerPlayer:
      type: object
      properties:
        name:
          type: string
          title: name
        provider:
          type: string
          title: provider
          description: 'The player''s identity at the game''s provider: ("steam", a SteamID64).'
        subject:
          type: string
          title: subject
        team:
          type: string
          title: team
        kills:
          type: integer
          title: kills
          format: int32
        deaths:
          type: integer
          title: deaths
          format: int32
        pingMs:
          type: integer
          title: ping_ms
          format: int32
        userId:
          type: string
          title: user_id
          description: The member that identity is linked to, resolved now; empty for a player who is not one.
      title: ServerPlayer
      additionalProperties: false
      description: ServerPlayer is one connected player at the last successful poll.
...
security: []
...